BugFix: listAllIdps() not show all Idps by filter when is called from listSavedIdps().
listSavedIdps() show saved IdPs. When defined filter not permit any saved IdP and result list of IdPs is empty then listSavedIdps() calls listAllIdps() and show all IdPs. In rare case when cached data is not available, list of IdPs wasn't complete.

is filter defined that list of IdPs is empty (saved IdPs is not permitted). In this case listSavedIdps() calls listAllIdps(). In